# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 伴AI API
  version: 1.27.0
  description: M1-M4 platform plus the M6 reliability control plane, governed Agent Studio, unified adjustable usage ledger, multi-instance runtime-configuration convergence, persisted cost and quality analytics with time trends, predictive budget burn and guarded model-switch recommendations, per-budget proactive forecast policies with sample gating, audited human recommendation feedback, revision-bound application and rollback-execution tracking, configurable before/after observation windows, downloadable periodic budget reviews, read-only rollback recommendations and post-rollback verification with acknowledged/closed human disposition, audited budget guardrails, automatic budget incidents with escalation plus email or management-console recovery notifications, baseline anomaly detection, Loki-backed redacted structured system logs with PostgreSQL fallback and Run/Trace correlation, threshold alert evaluation, downloadable incident evidence, durable W3C/OpenTelemetry Trace Context propagation across HTTP, Outbox, Kafka, Go/Python Agent workers and Langfuse, optional fail-open Langfuse LLM/Agent traces with generations, node observations and scores, operator DLQ replay, and compensation records. Every HTTP response includes canonical traceparent and X-Trace-ID correlation headers; clients may send traceparent/tracestate, while X-Trace-ID remains a normalized compatibility input.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/ops/alert-rules:
    get:
      operationId: listOperatorAlertRules
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      responses:
        "200":
          description: Persistent log-based alert rules.
          content:
            application/json:
              schema:
                type: object
                required: [rules]
                properties:
                  rules: { type: array, items: { $ref: "#/components/schemas/AlertRule" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createOperatorAlertRule
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateAlertRuleRequest" } } }
      responses:
        "201":
          description: Alert rule created with an audit record.
          content:
            application/json:
              schema:
                type: object
                required: [rule]
                properties:
                  rule: { $ref: "#/components/schemas/AlertRule" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /v1/ops/alert-subscriptions:
    get:
      operationId: listOperatorAlertSubscriptions
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      responses:
        "200":
          description: Admin-visible incident email routing subscriptions.
          content:
            application/json:
              schema:
                type: object
                required: [subscriptions]
                properties:
                  subscriptions: { type: array, items: { $ref: "#/components/schemas/AlertSubscription" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createOperatorAlertSubscription
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateAlertSubscriptionRequest" } } }
      responses:
        "201":
          description: Email subscription created and audited.
          content:
            application/json:
              schema:
                type: object
                required: [subscription]
                properties:
                  subscription: { $ref: "#/components/schemas/AlertSubscription" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /v1/ops/alerts/evaluate:
    post:
      operationId: evaluateOperatorAlerts
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      responses:
        "200":
          description: Immediate evaluation report. The API also evaluates enabled rules periodically.
          content:
            application/json:
              schema:
                type: object
                required: [evaluation]
                properties:
                  evaluation: { $ref: "#/components/schemas/AlertEvaluation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/ops/incidents:
    get:
      operationId: listOperatorIncidents
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      parameters:
        - { name: status, in: query, required: false, schema: { type: string, enum: [open, acknowledged, resolved] } }
        - { name: severity, in: query, required: false, schema: { type: string, enum: [warning, critical] } }
        - { name: source, in: query, required: false, schema: { type: string, enum: [log_rule, performance_budget] } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        "200":
          description: Incidents ordered by the most recent activity.
          content:
            application/json:
              schema:
                type: object
                required: [incidents]
                properties:
                  incidents: { type: array, items: { $ref: "#/components/schemas/Incident" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/incidents/{incident_id}:
    get:
      operationId: getOperatorIncident
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      parameters: [{ name: incident_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200":
          description: Incident detail including the rule snapshot and redacted evidence.
          content:
            application/json:
              schema:
                type: object
                required: [incident]
                properties:
                  incident: { $ref: "#/components/schemas/Incident" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/ops/incidents/{incident_id}/notifications:
    get:
      operationId: listOperatorIncidentNotifications
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      parameters: [{ name: incident_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200":
          description: Email delivery outcomes for incident opened, escalated, and resolved transitions. Recipient addresses are masked.
          content:
            application/json:
              schema:
                type: object
                required: [notifications]
                properties:
                  notifications: { type: array, items: { $ref: "#/components/schemas/IncidentNotification" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/ops/incidents/{incident_id}/evidence:
    get:
      operationId: exportOperatorIncidentEvidence
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      parameters:
        - { name: incident_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: format, in: query, required: false, schema: { type: string, enum: [json, markdown], default: json } }
      responses:
        "200":
          description: Privacy-reduced incident state, audit timeline, notification outcomes, and matching redacted logs.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IncidentEvidence" }
            text/markdown:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/incidents/{incident_id}/acknowledge:
    post:
      operationId: acknowledgeOperatorIncident
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      parameters: [{ name: incident_id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/IncidentStatusRequest" } } }
      responses:
        "200":
          description: Incident acknowledged with operator attribution and audit evidence.
          content:
            application/json:
              schema:
                type: object
                required: [incident]
                properties:
                  incident: { $ref: "#/components/schemas/Incident" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/incidents/{incident_id}/resolve:
    post:
      operationId: resolveOperatorIncident
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      parameters: [{ name: incident_id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/IncidentStatusRequest" } } }
      responses:
        "200":
          description: Incident manually resolved with operator attribution and audit evidence.
          content:
            application/json:
              schema:
                type: object
                required: [incident]
                properties:
                  incident: { $ref: "#/components/schemas/Incident" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /v1/mcp/servers:
    get:
      operationId: listMCPServers
      tags: [Skill]
      security: [{ bearerAuth: [] }]
      description: Lists operator-approved MCP server names and tool allowlists. Commands, arguments, and environments are never exposed.
      responses:
        "200":
          description: Configured stdio MCP allowlists; an empty list is the safe default.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/MCPServerSummary" } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    operatorBearerAuth:
      type: http
      scheme: bearer
      description: Uses either legacy OPERATOR_TOKEN or an operator account token. When operator MFA is enabled, clients must also send X-Operator-TOTP.
  responses:
    Unauthorized:
      description: Authentication is missing, invalid, expired, or revoked.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Resource already exists.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Authenticated operator role lacks permission for the requested action.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    ValidationError:
      description: Request failed domain validation.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource was not found in the caller scope.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    AlertRule:
      type: object
      additionalProperties: false
      required: [id, name, level, window_minutes, threshold, severity, enabled, created_by, updated_by, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        service: { type: string }
        level: { type: string, enum: [DEBUG, INFO, WARN, ERROR] }
        event_prefix: { type: string }
        window_minutes: { type: integer, minimum: 1, maximum: 1440 }
        threshold: { type: integer, minimum: 1, maximum: 1000000 }
        severity: { type: string, enum: [warning, critical] }
        enabled: { type: boolean }
        created_by: { type: string }
        updated_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateAlertRuleRequest:
      type: object
      additionalProperties: false
      required: [name, level, window_minutes, threshold, severity, enabled, reason]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        description: { type: string, maxLength: 512 }
        service: { type: string, maxLength: 128 }
        level: { type: string, enum: [DEBUG, INFO, WARN, ERROR] }
        event_prefix: { type: string, maxLength: 128 }
        window_minutes: { type: integer, minimum: 1, maximum: 1440 }
        threshold: { type: integer, minimum: 1, maximum: 1000000 }
        severity: { type: string, enum: [warning, critical] }
        enabled: { type: boolean }
        reason: { type: string, minLength: 1, maxLength: 512 }
...
    AlertEvaluation:
      type: object
      additionalProperties: false
      required: [rules_evaluated, triggered, opened, updated, resolved]
      properties:
        rules_evaluated: { type: integer, minimum: 0 }
        triggered: { type: integer, minimum: 0 }
        opened: { type: integer, minimum: 0 }
        updated: { type: integer, minimum: 0 }
        resolved: { type: integer, minimum: 0 }
    Incident:
      type: object
      additionalProperties: false
      required: [id, rule_name, source_type, source_id, status, severity, title, summary, level, observed_value, threshold, window_minutes, window_started_at, window_ended_at, evidence, opened_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        rule_id: { type: string, format: uuid }
        rule_name: { type: string }
        source_type: { type: string, enum: [log_rule, performance_budget] }
        source_id: { type: string, format: uuid }
        status: { type: string, enum: [open, acknowledged, resolved] }
        severity: { type: string, enum: [warning, critical] }
        title: { type: string }
        summary: { type: string }
        service: { type: string }
        level: { type: string, enum: [DEBUG, INFO, WARN, ERROR] }
        event_prefix: { type: string }
        observed_value: { type: integer, minimum: 0 }
        threshold: { type: integer, minimum: 1 }
        window_minutes: { type: integer, minimum: 1 }
        window_started_at: { type: string, format: date-time }
        window_ended_at: { type: string, format: date-time }
        evidence: { type: object, additionalProperties: true }
        opened_at: { type: string, format: date-time }
        acknowledged_at: { type: string, format: date-time }
        acknowledged_by: { type: string }
        resolved_at: { type: string, format: date-time }
        resolved_by: { type: string }
        resolution: { type: string }
        updated_at: { type: string, format: date-time }
    IncidentStatusRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 1, maxLength: 512 }
    AlertSubscription:
      type: object
      additionalProperties: false
      required: [id, name, channel, target, minimum_severity, notify_on_open, notify_on_resolved, enabled, created_by, updated_by, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        rule_id: { type: string, format: uuid }
        rule_name: { type: string }
        channel: { type: string, enum: [email, console] }
        target: { type: string, description: Email address for email or a stable management-console inbox key. }
        minimum_severity: { type: string, enum: [warning, critical] }
        notify_on_open: { type: boolean }
        notify_on_resolved: { type: boolean }
        enabled: { type: boolean }
        created_by: { type: string }
        updated_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateAlertSubscriptionRequest:
      type: object
      additionalProperties: false
      required: [name, channel, target, minimum_severity, notify_on_open, notify_on_resolved, enabled, reason]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        rule_id: { type: string, format: uuid }
        channel: { type: string, enum: [email, console] }
        target: { type: string, maxLength: 320 }
        minimum_severity: { type: string, enum: [warning, critical] }
        notify_on_open: { type: boolean }
        notify_on_resolved: { type: boolean }
        enabled: { type: boolean }
        reason: { type: string, minLength: 1, maxLength: 512 }
...
    IncidentNotification:
      type: object
      additionalProperties: false
      required: [id, transition, channel, target, status, attempts, created_at]
      properties:
        id: { type: string, format: uuid }
        transition: { type: string, enum: [opened, escalated, resolved] }
        channel: { type: string, enum: [email, console] }
        target: { type: string, description: Masked recipient address for email or visible inbox key for management-console delivery. }
        delivery_id: { type: string, format: uuid, description: Present only for email delivery. }
        status: { type: string, enum: [queued, processing, sent, failed] }
        attempts: { type: integer, minimum: 0 }
        failure_code: { type: string }
        created_at: { type: string, format: date-time }
        sent_at: { type: string, format: date-time }
    IncidentActivity:
      type: object
      additionalProperties: false
      required: [action, actor_type, actor, occurred_at]
      properties:
        action: { type: string }
        actor_type: { type: string, enum: [operator, system] }
        actor: { type: string }
        reason: { type: string }
        occurred_at: { type: string, format: date-time }
    IncidentEvidenceLog:
      type: object
      additionalProperties: false
      required: [id, occurred_at, service, level, event, message]
      properties:
        id: { type: integer, minimum: 1 }
        occurred_at: { type: string, format: date-time }
        service: { type: string }
        level: { type: string, enum: [DEBUG, INFO, WARN, ERROR] }
        event: { type: string }
        message: { type: string }
        trace_id: { type: string }
        run_id: { type: string }
        error_code: { type: string }
    IncidentEvidence:
      type: object
      additionalProperties: false
      required: [incident, activity, notifications, logs, generated_at]
      properties:
        incident: { $ref: "#/components/schemas/Incident" }
        activity: { type: array, items: { $ref: "#/components/schemas/IncidentActivity" } }
        notifications: { type: array, items: { $ref: "#/components/schemas/IncidentNotification" } }
        logs: { type: array, items: { $ref: "#/components/schemas/IncidentEvidenceLog" } }
        generated_at: { type: string, format: date-time }
...
    MCPServerSummary:
      type: object
      additionalProperties: false
      required: [name, enabled, allowed_tools, protocol_version, transport]
      properties:
        name: { type: string }
        enabled: { type: boolean }
        allowed_tools: { type: array, items: { type: string } }
        protocol_version: { type: string, enum: ["2025-11-25"] }
        transport: { type: string, enum: [stdio] }
...